+++ ts/features/idpay/details/hooks/useInitiativeTimelineFetcher.tsx
@@ -1,16 +1,6 @@
-import { Divider, ListItemNav, VSpacer } from "@pagopa/io-app-design-system";
 import * as pot from "@pagopa/ts-commons/lib/pot";
-import { useNavigation } from "@react-navigation/native";
 import * as React from "react";
-import I18n from "../../../../i18n";
-import {
-  AppParamsList,
-  IOStackNavigationProp
-} from "../../../../navigation/params/AppParamsList";
 import { useIODispatch, useIOSelector } from "../../../../store/hooks";
-import { useIOBottomSheetAutoresizableModal } from "../../../../utils/hooks/bottomSheet";
-import { idPayGenerateBarcode } from "../../barcode/store/actions";
-import { IDPayPaymentRoutes } from "../../payment/navigation/navigator";
 import {
   idpayOperationListSelector,
   idpayPaginatedTimelineSelector,
@@ -91,54 +81,4 @@ export const useInitiativeTimelineFetcher = (
     currentPage,
     lastUpdate
   } as const;
-};
-
-export const useIdpayDiscountDetailsBottomSheet = (initiativeId: string) => {
-  const navigation = useNavigation<IOStackNavigationProp<AppParamsList>>();
-  const navigateToPaymentAuthorization = () => {
-    navigation.navigate(IDPayPaymentRoutes.IDPAY_PAYMENT_CODE_SCAN);
-  };
-  const dispatch = useIODispatch();
-  const barcodePressHandler = () => {
-    dispatch(idPayGenerateBarcode.request({ initiativeId }));
-    bottomSheet.dismiss();
-  };
-
-  const DiscountInitiativeBottomSheetContent = () => (
-    <>
-      <ListItemNav
-        value={I18n.t(
-          "idpay.initiative.discountDetails.bottomSheetOptions.scanQr"
-        )}
-        icon="qrCode"
-        onPress={() => {
-          bottomSheet.dismiss();
-          navigateToPaymentAuthorization();
-        }}
-        accessibilityLabel={I18n.t(
-          "idpay.initiative.discountDetails.bottomSheetOptions.scanQr"
-        )}
-      />
-      <Divider />
-      <ListItemNav
-        icon="barcode"
-        value={I18n.t(
-          "idpay.initiative.discountDetails.bottomSheetOptions.generateBarcode"
-        )}
-        onPress={barcodePressHandler}
-        accessibilityLabel={I18n.t(
-          "idpay.initiative.discountDetails.bottomSheetOptions.generateBarcode"
-        )}
-      />
-      <Divider />
-      <VSpacer size={24} />
-    </>
-  );
-
-  const bottomSheet = useIOBottomSheetAutoresizableModal({
-    component: <DiscountInitiativeBottomSheetContent />,
-    title: null
-  });
-
-  return bottomSheet;
 };
