+++ ts/features/idpay/details/screens/OperationsListScreen.tsx
@@ -22,7 +22,7 @@ import { showToast } from "../../../../utils/showToast";
 import { useTimelineDetailsBottomSheet } from "../../timeline/components/TimelineDetailsBottomSheet";
 import { TimelineOperationListItem } from "../components/TimelineOperationListItem";
 import { IDPayDetailsParamsList } from "../navigation";
-import { useInitiativeTimelineFetcher } from "../utils/hooks";
+import { useInitiativeTimelineFetcher } from "../hooks/useInitiativeTimelineFetcher";
 import { localeDateFormat } from "../../../../utils/locale";
 export type OperationsListScreenParams = { initiativeId: string };
